[web] Extract the WiFi connection form
