+++ web/src/WirelessSelector.jsx
@@ -21,18 +21,12 @@
 
 import React, { useState } from "react";
 import {
-  Alert,
   Button,
   Card,
   CardBody,
-  Form,
-  FormGroup,
-  FormSelect,
-  FormSelectOption,
   Radio,
   Split,
   SplitItem,
-  TextInput
 } from "@patternfly/react-core";
 
 import {
@@ -43,95 +37,13 @@ import {
 import Popup from "./Popup";
 import Center from "./Center";
 import WifiNetworkMenu from "./WifiNetworkMenu";
+import WifiConnectionForm from "./WifiConnectionForm";
 
-import { useInstallerClient } from "./context/installer";
-
-const CONNECTION_FORM_ID = "chosen-network";
 const baseHiddenNetwork = { ssid: "", hidden: true };
 
-function WirelessConnectionForm({ network, setSubmittingData, onClose }) {
-  const client = useInstallerClient();
-  const [error, setError] = useState(false);
-  const [ssid, setSsid] = useState(network?.ssid);
-  const [password, setPassword] = useState("");
-  const [security, setSecurity] = useState("none");
-
-  const security_options = [
-    { value: "", label: "None" },
-    { value: "wpa-psk", label: "WPA & WPA2 Personal" },
-    { value: "wpa-eap", label: "WPA & WPA2 Enterprise" }
-  ];
-
-  const selectorOptions = security_options.map(security => (
-    <FormSelectOption key={security.value} value={security.value} label={security.label} />
-  ));
-
-  const connectNetwork = async () => {
-    await client.network.addAndConnectTo(ssid, { security, password });
-  };
-
-  const accept = async e => {
-    e.preventDefault();
-    setError(false);
-    setSubmittingData(true);
-
-    connectNetwork()
-      .then(() => onClose())
-      .catch(() => {
-        setError(true);
-        setSubmittingData(false);
-      });
-  };
-
-  return (
-    <Form id={CONNECTION_FORM_ID} onSubmit={accept}>
-      { error &&
-        <Alert variant="warning" isInline title="Something went wrong">
-          <p>Please, review provided settings and try again.</p>
-        </Alert> }
-
-      { network?.hidden &&
-        <FormGroup fieldId="ssid" label="SSID">
-          <TextInput
-            id="ssid"
-            name="ssid"
-            label="SSID"
-            aria-label="ssid"
-            value={ssid}
-            onChange={setSsid}
-          />
-        </FormGroup> }
-
-      <FormGroup fieldId="security" label="Security">
-        <FormSelect
-          id="security"
-          aria-label="security"
-          value={security}
-          onChange={setSecurity}
-        >
-          {selectorOptions}
-        </FormSelect>
-      </FormGroup>
-      { security === "wpa-psk" &&
-        <FormGroup fieldId="password" label="WPA Password">
-          <TextInput
-            id="password"
-            name="password"
-            aria-label="Password"
-            value={password}
-            label="Password"
-            onChange={setPassword}
-          />
-        </FormGroup> }
-    </Form>
-  );
-}
-
 function WirelessSelector({ activeConnections, connections, accessPoints, onClose }) {
-  const client = useInstallerClient();
   const [selected, setSelected] = useState(null);
-  const [isLoading, setIsLoading] = useState(false);
-  const [actionsDisabled, setActionsDisabled] = useState(false);
+  const unsetSelected = () => setSelected(null);
 
   const networks = accessPoints.sort((a, b) => b.strength - a.strength);
   const ssids = networks.map(a => a.ssid);
@@ -139,19 +51,8 @@ function WirelessSelector({ activeConnections, connections, accessPoints, onClos
     return (ap.ssid !== "" && ssids.indexOf(ap.ssid) >= index);
   });
 
-  const setSubmittingData = (value) => {
-    setIsLoading(value);
-    setActionsDisabled(value);
-  };
-
   const isSelected = (network) => selected === network.ssid;
   const isConnected = (network) => activeConnections.find((n) => n.name === network.ssid);
-
-  const deleteConnection = (ssid) => {
-    const conn = connections.find((n) => n.name === ssid);
-
-    client.network.deleteConnection(conn);
-  };
 
   const renderFilteredNetworks = () => {
     return filtered.map(n => {
@@ -193,11 +94,7 @@ function WirelessSelector({ activeConnections, connections, accessPoints, onClos
             { selected &&
               <Split hasGutter>
                 <SplitItem isFilled className="content">
-                  <WirelessConnectionForm
-                    network={n}
-                    setSubmittingData={setSubmittingData}
-                    onClose={onClose}
-                  />
+                  <WifiConnectionForm network={n} onCancel={unsetSelected} />
                 </SplitItem>
               </Split> }
           </CardBody>
@@ -213,19 +110,19 @@ function WirelessSelector({ activeConnections, connections, accessPoints, onClos
           <CardBody>
             <Split hasGutter className="content">
               <SplitItem isFilled>
-                <WirelessConnectionForm
-                  network={selected}
-                  setSubmittingData={setSubmittingData}
-                  onClose={onClose}
-                />
-                <Button variant="link" onClick={() => setSelected(null)}>Cancel</Button>
+                <WifiConnectionForm network={selected} onCancel={unsetSelected} />
               </SplitItem>
             </Split>
           </CardBody>
         </Card>
         { !selected?.hidden &&
           <Center>
-            <Button variant="link" onClick={() => setSelected(baseHiddenNetwork)}>Add network manually</Button>
+            <Button
+              variant="link"
+              onClick={() => setSelected(baseHiddenNetwork)}
+            >
+              Add network manually
+            </Button>
           </Center> }
       </>
     );
@@ -237,15 +134,7 @@ function WirelessSelector({ activeConnections, connections, accessPoints, onClos
       { renderHiddenNetworkForm() }
 
       <Popup.Actions>
-        <Popup.Confirm
-          type="submit"
-          form={CONNECTION_FORM_ID}
-          isDisabled={actionsDisabled || !selected}
-          isLoading={isLoading}
-        >
-          { isLoading ? "Connecting" : "Connect" }
-        </Popup.Confirm>
-        <Popup.Cancel onClick={onClose} isDisabled={actionsDisabled} />
+        <Popup.PrimaryAction onClick={onClose}>Close</Popup.PrimaryAction>
       </Popup.Actions>
     </Popup>
   );
