+++ web/src/patternfly.scss
@@ -104,9 +104,7 @@
   grid-column: 1/3;
 }
 
-.pf-c-modal-box__footer {
-}
-
+.pf-c-form__actions,
 .pf-c-modal-box__footer {
   // EOS prefers buttons placed at the right
   // Read https://eosdesignsystem.herokuapp.com/buttons/positioning
